--- packages/matrix-util/src/vec3.ts
@@ -1,1 +1,0 @@
-export * from 'gl-matrix/vec3';
